+++ packages/storefront/src/app/(main)/components/link/link.meta.ts
@@ -1,5 +1,10 @@
 import { componentMeta } from '@porsche-design-system/component-meta';
-import Accessibility from '@/app/(main)/components/link/accessibility/page.mdx';
+import { ariaOnComponentHostVsAriaPropA11yExample } from '@/app/(main)/components/link/accessibility/examples/aria-on-component-host-vs-aria-prop/example';
+import { iconOnlyLinkWithoutAccessibleNameA11yExample } from '@/app/(main)/components/link/accessibility/examples/icon-only-link-without-accessible-name/example';
+import { indicatingTheCurrentPageA11yExample } from '@/app/(main)/components/link/accessibility/examples/indicating-the-current-page/example';
+import { vagueLinkLabelWithoutContextA11yExample } from '@/app/(main)/components/link/accessibility/examples/vague-link-label-without-context/example';
+import AccessibilityOverview from '@/app/(main)/components/link/accessibility/overview.mdx';
+import AccessibilityTests from '@/app/(main)/components/link/accessibility/tests.mdx';
 import IntroductionDescription from '@/app/(main)/components/link/configurator/introduction.mdx';
 import { linkSlotStories, linkStory } from '@/app/(main)/components/link/configurator/story';
 import FrameworkRoutingDescription from '@/app/(main)/components/link/examples/framework-routing/example.mdx';
@@ -30,7 +35,15 @@ export const linkMeta = {
     },
   },
   usage: Usage,
-  accessibility: Accessibility,
+  accessibility: {
+    overview: AccessibilityOverview,
+    examples: {
+      ariaOnComponentHostVsAriaProp: ariaOnComponentHostVsAriaPropA11yExample,
+      iconOnlyLinkWithoutAccessibleName: iconOnlyLinkWithoutAccessibleNameA11yExample,
+      indicatingTheCurrentPage: indicatingTheCurrentPageA11yExample,
+      vagueLinkLabelWithoutContext: vagueLinkLabelWithoutContextA11yExample,
+    },
+    tests: AccessibilityTests,
+  },
   api: componentMeta['p-link'],
 } satisfies ComponentDocsMeta<'p-link'>;
-