+++ packages/storefront/src/app/(main)/components/multi-select/multi-select.meta.ts
@@ -5,7 +5,11 @@ import {
   multiSelectExampleForm,
   multiSelectExampleSelectedSlot,
 } from '@porsche-design-system/shared/examples';
-import Accessibility from '@/app/(main)/components/multi-select/accessibility/page.mdx';
+import { hiddenLabelWithoutAccessibleNameA11yExample } from '@/app/(main)/components/multi-select/accessibility/examples/hidden-label-without-accessible-name/example';
+import { promptOptionUsedInsteadOfALabelA11yExample } from '@/app/(main)/components/multi-select/accessibility/examples/prompt-option-used-instead-of-a-label/example';
+import { validationFeedbackViaStateAndMessageApiA11yExample } from '@/app/(main)/components/multi-select/accessibility/examples/validation-feedback-via-state-and-message-api/example';
+import AccessibilityOverview from '@/app/(main)/components/multi-select/accessibility/overview.mdx';
+import AccessibilityTests from '@/app/(main)/components/multi-select/accessibility/tests.mdx';
 import IntroductionDescription from '@/app/(main)/components/multi-select/configurator/introduction.mdx';
 import { multiSelectSlotStories, multiSelectStory } from '@/app/(main)/components/multi-select/configurator/story';
 import AsyncFilterDescription from '@/app/(main)/components/multi-select/examples/async-filter/example.mdx';
@@ -64,7 +68,14 @@ export const multiSelectMeta = {
     },
   },
   usage: Usage,
-  accessibility: Accessibility,
+  accessibility: {
+    overview: AccessibilityOverview,
+    examples: {
+      promptOptionUsedInsteadOfALabel: promptOptionUsedInsteadOfALabelA11yExample,
+      hiddenLabelWithoutAccessibleName: hiddenLabelWithoutAccessibleNameA11yExample,
+      validationFeedbackViaStateAndMessageApi: validationFeedbackViaStateAndMessageApiA11yExample,
+    },
+    tests: AccessibilityTests,
+  },
   api: componentMeta['p-multi-select'],
 } satisfies ComponentDocsMeta<'p-multi-select'>;
-